feat: add DEBUG_PKG diagnostics to SEA mode, deduplicate diagnostic code
Move VFS tree dump and fs call tracing from prelude/diagnostic.js into
bootstrap-shared.js as installDiagnostic(). Both bootstraps now share
the same diagnostic implementation.

Security: diagnostics are only available when built with --debug / -d.
In SEA mode this is controlled via manifest.debug flag set at build
time; without it, installDiagnostic is never called.

- Delete prelude/diagnostic.js (replaced by shared installDiagnostic)
- Packer injects small DICT-dump snippet + shared call when --debug
- SEA bootstrap gates on manifest.debug before calling installDiagnostic
- Fix assertSeaNodeVersion: restore check to Node >= 20 (simple SEA)
- Fix withSeaTmpDir: restore tmpDir cleanup (was commented out)
- Update architecture docs with diagnostic usage
