+++ prelude/sea-bootstrap.js
@@ -179,6 +179,16 @@ shared.patchChildProcess(entrypoint);
 shared.setupProcessPkg(entrypoint);
 
 // /////////////////////////////////////////////////////////////////
+// DIAGNOSTICS /////////////////////////////////////////////////////
+// /////////////////////////////////////////////////////////////////
+
+// Only available when the binary was built with --debug / -d.
+// At runtime, set DEBUG_PKG=1 (VFS tree) or DEBUG_PKG=2 (+ fs tracing).
+if (manifest.debug) {
+  shared.installDiagnostic(SNAPSHOT_PREFIX);
+}
+
+// /////////////////////////////////////////////////////////////////
 // ENTRYPOINT //////////////////////////////////////////////////////
 // /////////////////////////////////////////////////////////////////
